Reworking Checkout Process
Reworking checkout process for multiple active gateways.

Building a framework for all the gateways to implement, starting with
Stripe gateway as the model.

Stripe is working partially (without trials) the other gateways still
needs to be converted.

Updated Stripe PHP API to 1.18 (last version to support PHP 5.2, but
compatible with new API updates '2015-02-16').

Fixing number of issues with webhooks and syncing plans with Stripe.